+++ src/main/java/com/danilore/piniateria_lizzety/service/persona/PersonaService.java
@@ -1,11 +1,12 @@
 package com.danilore.piniateria_lizzety.service.persona;
 
+import com.danilore.piniateria_lizzety.dto.persona.PersonaDTO;
 import com.danilore.piniateria_lizzety.exception.DAOException;
 import com.danilore.piniateria_lizzety.model.persona.Persona;
 import com.danilore.piniateria_lizzety.repository.persona.PersonaRepository;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
@@ -17,30 +18,40 @@ public class PersonaService {
     @Autowired
     private PersonaRepository personaRepository;
 
-    public Page<Persona> listarTodas(Pageable pageable) {
-        return personaRepository.findAll(pageable);
+    public Page<PersonaDTO> getAll(int page, int size) {
+        Pageable pageable = PageRequest.of(page, size);
+        Page<Persona> personaPage = personaRepository.findAll(pageable);
+
+        return personaPage.map(PersonaDTO::fromEntity);
     }
 
-    public Persona buscarPorId(Long id) {
-        return personaRepository.findById(id)
-                .orElseThrow(() -> new DAOException("La Persona con ID: " + id + " no existe"));
+    public PersonaDTO getById(Long id) {
+        Persona persona = personaRepository.findById(id)
+                .orElseThrow(() -> new IllegalArgumentException("Persona no encontrado"));
+        return PersonaDTO.fromEntity(persona);
     }
 
-    public Persona guardar(Persona persona) {
+    public PersonaDTO save(PersonaDTO personaDTO) {
+        Persona persona = personaDTO.toEntity();
+
         if (personaRepository.findByCorreo(persona.getCorreo()).isPresent()) {
             throw new DAOException("El correo ya está registrado.");
         }
         if (personaRepository.findByTelefono(persona.getTelefono()).isPresent()) {
-            throw new DAOException("La persona ya está registrada.");
+            throw new DAOException("El telefono ya está registrado.");
 
         }
 
-        return personaRepository.save(persona);
+        Persona savedPersona = personaRepository.save(persona);
+        return PersonaDTO.fromEntity(savedPersona);
     }
 
-    public Persona editar(Long id, Persona personaActualizada) {
+    public PersonaDTO update(Long id, PersonaDTO personaDTO) {
+        // Convertir el DTO en una entidad
+        Persona personaActualizada = personaDTO.toEntity();
+
         Persona personaExistente = personaRepository.findById(id)
-                .orElseThrow(() -> new DAOException("Persona no encontrado"));
+                .orElseThrow(() -> new DAOException("Persona no encontrado con ID: " + id));
 
         personaExistente.setNombres(personaActualizada.getNombres());
         personaExistente.setApellidos(personaActualizada.getApellidos());
@@ -55,16 +66,28 @@ public class PersonaService {
         personaExistente.setFechaNacimiento(personaActualizada.getFechaNacimiento());
         personaExistente.setGenero(personaActualizada.getGenero());
         personaExistente.setLugarNacimiento(personaActualizada.getLugarNacimiento());
-        return personaRepository.save(personaExistente);
+
+        // Convertir la entidad actualizada nuevamente a DTO y devolverla
+        return PersonaDTO.fromEntity(personaRepository.save(personaExistente));
     }
 
-
-
-    public void eliminarPorId(Long id) {
+    public void deleteById(Long id) {
+        if (!personaRepository.existsById(id)) {
+            throw new DAOException("Persona no encontrado");
+        }
         personaRepository.deleteById(id);
     }
 
-    // Métodos personalizados
+    // Listar usuario por nombres, apellidos, numeroDocumento, correo o ID
+    public Page<PersonaDTO> buscarPorCriterio(String criterio, int page, int size) {
+        if (criterio == null || criterio.isBlank()) {
+            throw new IllegalArgumentException("El criterio de búsqueda no puede estar vacío");
+        }
+        Pageable pageable = PageRequest.of(page, size, Sort.by("apellidos").ascending());
+        Page<Persona> personaPage = personaRepository.buscarPorCriterio(criterio, pageable);
+        return personaPage.map(PersonaDTO::fromEntity);
+    }
+
     public Persona BuscarPorIdConTipoDocumento(Long id) {
         return personaRepository.findByIdConTipoDocumento(id)
                 .orElseThrow(() -> new DAOException("Persona no encontrado"));
